navisight backend (2)/navisightbackend/backend/requirements.txt
transformers
torch
ultralytics
pillow
matplotlib
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attr==0.3.1
attrs==23.2.0
audioread==3.0.1
azure-core==1.32.0
azure-storage-blob==12.24.1
bcrypt==4.1.3
black==24.8.0
bleach==5.0.1
blinker==1.8.2
boto==2.49.0
boto3==1.37.6
botocore==1.37.6
cachetools==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
datamodel-code-generator==0.26.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
Django==5.1.6
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.6.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
einops==0.8.0
email_validator==2.2.0
executing==2.0.1
expiringdict==1.2.2
Faker==36.1.1
fastapi==0.115.12
filelock==3.14.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
g==0.0.4
gast==0.5.4
genson==1.3.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.2
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
greenlet==3.0.3
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.23.3
humansignal-drf-yasg==1.21.10.post1
idna==3.7
ijson==3.3.0
imageio==2.34.1
importlib_metadata==8.5.0
inflect==5.6.2
inflection==0.5.1
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.3.3
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
label-studio==1.16.0
label-studio-sdk==1.0.10
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lockfile==0.12.2
logger==1.4
lxml==5.3.1
lxml_html_clean==0.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.5
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.65.3
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.82
opentelemetry-api==1.30.0
openvino==2024.6.0
openvino-dev==2024.6.0
openvino-telemetry==2025.1.0
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.0.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.6
pooch==1.8.2
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf==4.25.7
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyboxen==1.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygeoip==0.3.2
PyGetWindow==0.0.9
Pygments==2.17.2
pymongo==4.7.2
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
pyRFC3339==2.0.1
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.4
pytweening==1.2.0
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.23.1
rq==1.16.2
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.11.4
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.22.0
setuptools==75.6.0
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.15
tk==0.1.0
tldextract==5.1.3
tokenizers==0.19.1
toml==0.10.2
torch==2.5.1
torchdiffeq==0.2.4
torchsde==0.2.6
torchvision==0.20.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
transformers==4.41.2
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.78
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.34.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.3.8
wheel==0.40.0
wrapt==1.16.0
xmljson==0.2.1
yarl==1.9.4
zipp==3.21.0